+++ src/ecs/em-systems.ts
@@ -5,9 +5,9 @@ import {
   _em,
   Entities,
   EntityW,
-  isDeadC,
   ReadonlyEntities,
 } from "./entity-manager.js";
+import { isDeadC } from "./em-components.js";
 import { ComponentDef } from "./em-components.js";
 import { DBG_SYSTEM_ORDER, DBG_INIT_CAUSATION } from "../flags.js";
 import { resetTempMatrixBuffer } from "../matrix/sprig-matrix.js";
