+++ src/net/events.ts
@@ -1,4 +1,6 @@
-import { EM, EDef, ESet, Entity } from "../ecs/entity-manager.js";
+import { EM, Entity } from "../ecs/entity-manager.js";
+import { ESet } from "../ecs/em-components.js";
+import { EDef } from "../ecs/em-components.js";
 import { Component } from "../ecs/em-components.js";
 import { ComponentDef } from "../ecs/em-components.js";
 import {
